requirements.txt
fastapi
uvicorn
pydantic
python-multipart
python-jose
passlib[bcrypt]
supabase
openai>=1.40.0
python-docx
pypdf
tiktoken